Make portRe constant
When running `unfurl format %r.%t` on a long list of urls, the function `extractFromDomain()` will be called separately for each one of the urls, multiple times. 
This causes the same regex to be compiled multiple times:
```
	portRe := regexp.MustCompile(`(?m):\d+$`)
```

Instead, portRe should be a constant. I changed the regexp to be constant and tested it on a list of 1.1 Million urls. I got the following results:

Old: 0m35.384s
New: 0m21.374s